requirements.txt
PyYAML~=5.4.1
countrycode~=0.2